Implemented DB Helpers
2-DB Structure (Excursion db and Picture db: Excursions contain list of
locations to display old trace, name, etc. Pictures have a bitmap,
location, and row ID of an associated Excursion)

To save pictures, we have to first make and save a new Excursion to
generate a unique ID to tag Pictures taken subsequently during that
excursion. Once the Excursion is complete, it can be updated (using the
updateEntry db helper function) to include more stuff)